remove combine daterangepicker
